refactor(workflow_runner): extract CollaborationCoordinator and AgentRouter collaborators (#6393 #6392)
WorkflowRunner had three distinct responsibilities: execution, agent routing,
and Redis pub/sub coordination. This commit separates them into focused classes:

- CollaborationCoordinator (enhanced_orchestration/collaboration_coordinator.py):
  owns _ensure_redis, coordinate_collaboration, broadcast_to_agents

- AgentRouter (enhanced_orchestration/agent_router.py):
  owns get_agent_recommendations, get_agent_instance, calculate_capability_coverage.
  Takes agent_registry as a constructor parameter (#6392 — fixes hidden construction)

- WorkflowRunner: now pure execution engine (~165 lines, down from ~278).
  Delegates to injected CollaborationCoordinator and AgentRouter.

- Orchestrator._init_strategy_components: constructs AgentClientRegistry,
  CollaborationCoordinator, and AgentRouter; injects all into WorkflowRunner.
