+++ autobot-backend/enhanced_orchestration/__init__.py
@@ -14,8 +14,12 @@ Sub-modules:
 - execution_strategies.py: Strategy implementations (sequential, parallel, pipeline, etc.)
 - workflow_planning.py: Workflow planning, building, and utilities
 - success_criteria.py: Structured success criteria evaluation
+- collaboration_coordinator.py: Redis pub/sub collaboration layer (#6393)
+- agent_router.py: Agent selection, resolution, capability coverage (#6393/#6392)
 """
 
+from .agent_router import AgentRouter
+from .collaboration_coordinator import CollaborationCoordinator
 from .execution_strategies import ExecutionStrategyHandler
 from .success_criteria import (
     CriteriaResult,
@@ -58,4 +62,7 @@ __all__ = [
     "SuccessCriteriaEvaluator",
     # Execution engine (#5058)
     "WorkflowRunner",
+    # Collaborators extracted from WorkflowRunner (#6393/#6392)
+    "AgentRouter",
+    "CollaborationCoordinator",
 ]
